Simple Linked List
This is an implementation of SLL, it only contains one head and each node contains the address of next node but not the previous, that is why it is also called Singly Linked List. The end of this link list is identified by the node that has null address, meaning no address for the next node.